fix/course_dtos :anotation validation fix

+++ src/main/java/com/example/e_learning_system/Repository/UserRepository.java
@@ -15,10 +15,10 @@ import java.util.Optional;
 
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 
     public Optional<UserEntity> findByEmail(String email);
-    public Optional<UserEntity> findById(long id);
+    public Optional<UserEntity> findById(Integer id);
 
 
     public Page<UserEntity> findAll(Pageable pageable);
